Create meeting_2_Notes
Notes from our second meeting. 

We defined our roles further, and everyone picked a page to do.

We discussed font colour, font size, colour schemes, etc. But we did not come to a solid conclusion on those. The notes contain the fonts Giri has chosen/suggested. 

We agreed on what we would do before the next meeting (3) and the notes contain the specific time we will meet.